requirements.txt
scikit-image==0.17.1
numpy==1.16.4
tensorflow-cpu==2.2.0
matplotlib